+++ src/main/java/zcore/config/Config.java
@@ -1,0 +1,15 @@
+package zcore.config;
+
+import net.minecraftforge.common.config.Configuration;
+
+/**
+ * A place to add all your config options
+ * @author Z_Doctor
+ */
+public class Config {
+	public static void load(Configuration c) {
+		//Example: propertyName = c.get(category, key, defaultValue);
+	}
+	//Example: public static Property propertyName;
+	
+}