+++ src/main/java/zcore/events/CustomEvents.java
@@ -1,0 +1,17 @@
+package zcore.events;
+
+import zcore.common.EasyStuff.EventRegistry;
+
+public class CustomEvents {
+	/**
+	 * Place all your FML Events here. Remember to annotate with @SubscribeEvent
+	 * @author Z_Doctor
+	 */
+	public static class FML {
+		
+	}
+	
+	public static class Forge {
+		
+	}
+}