+++ src/main/java/zcore/gameregistry/ZBlocks.java
@@ -1,0 +1,8 @@
+package zcore.gameregistry;
+/**
+ * A class to hold all your Blocks. Be sure to properly assign them.
+ * @author Z_Doctor
+ */
+public class ZBlocks {
+	// Example: public static Block blockName;
+}