+++ src/main/java/zcore/gameregistry/ZItems.java
@@ -1,0 +1,8 @@
+package zcore.gameregistry;
+/**
+ * A class to hold all your Items. Be sure to properly assign them.
+ * @author Z_Doctor
+ */
+public class ZItems {
+	// Example: public static Item itemName;
+}